Decompose own descriptor lookup by receiver and constructor
